move tests for programming-exercise.utils.ts into a single location
